+++ rdr/OutStream.h
@@ -95,7 +95,7 @@ namespace rdr {
       const U8* dataPtr = (const U8*)data;
       const U8* dataEnd = dataPtr + length;
       while (dataPtr < dataEnd) {
-        int n = check(1, dataEnd - dataPtr);
+        int n = check(1, (int)(dataEnd - dataPtr));
         memcpy(ptr, dataPtr, n);
         ptr += n;
         dataPtr += n;
